+++ ThirdPartyAdapters/moloco/moloco/src/test/kotlin/com/google/ads/mediation/moloco/MolocoMediationAdapterTest.kt
@@ -29,6 +29,7 @@ import com.google.ads.mediation.moloco.MolocoMediationAdapter.Companion.MEDIATIO
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.AdFormat
 import com.google.android.gms.ads.AdSize
+import com.google.android.gms.ads.AgeRestrictedTreatment
 import com.google.android.gms.ads.MobileAds
 import com.google.android.gms.ads.RequestConfiguration
 import com.google.android.gms.ads.mediation.InitializationCompleteCallback
@@ -117,6 +118,15 @@ class MolocoMediationAdapterTest {
   @Before
   fun setUp() {
     adapter = MolocoMediationAdapter()
+    val requestConfig =
+      RequestConfiguration.Builder()
+        .setTagForChildDirectedTreatment(
+          RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
+        )
+        .setTagForUnderAgeOfConsent(RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_UNSPECIFIED)
+        .setAgeRestrictedTreatment(AgeRestrictedTreatment.UNSPECIFIED)
+        .build()
+    MobileAds.setRequestConfiguration(requestConfig)
   }
 
   // region Version Tests
@@ -240,6 +250,36 @@ class MolocoMediationAdapterTest {
             .setTagForChildDirectedTreatment(
               RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE
             )
+            .build()
+        MobileAds.setRequestConfiguration(requestConfig)
+
+        adapter.initialize(
+          context,
+          mockInitializationCompleteCallback,
+          listOf(mediationConfiguration),
+        )
+
+        mockedMoloco.verify({ initialize(any(), molocoCallbackCaptor.capture()) }, times(1))
+        val molocoCallback = molocoCallbackCaptor.firstValue
+        molocoCallback.onMolocoInitializationStatus(
+          MolocoInitStatus(Initialization.SUCCESS, "Test")
+        )
+        mockedMolocoUtils.verify { setMolocoIsAgeRestricted(true) }
+        verify(mockInitializationCompleteCallback).onInitializationSucceeded()
+      }
+    }
+  }
+
+  @Test
+  fun initialize_withAgeRestrictedTreatmentChild_configuresPrivacyTrue() {
+    mockStatic(MolocoAdapterUtils::class.java).use { mockedMolocoUtils ->
+      mockStatic(Moloco::class.java).use { mockedMoloco ->
+        val serverParameters = bundleOf(MolocoMediationAdapter.KEY_APP_KEY to TEST_APP_KEY_1)
+        val mediationConfiguration = createMediationConfiguration(AdFormat.BANNER, serverParameters)
+        val molocoCallbackCaptor = argumentCaptor<MolocoInitializationListener>()
+        val requestConfig =
+          RequestConfiguration.Builder()
+            .setAgeRestrictedTreatment(AgeRestrictedTreatment.CHILD)
             .build()
         MobileAds.setRequestConfiguration(requestConfig)